tidy tracing decorators
