+++ clients/builderapi/tracing.go
@@ -10,13 +10,13 @@ import (
 	"github.com/opentracing/opentracing-go/log"
 )
 
-type tracingClient struct {
-	Client
-}
-
 // NewTracingClient returns a new instance of a tracing Client.
 func NewTracingClient(c Client) Client {
 	return &tracingClient{c}
+}
+
+type tracingClient struct {
+	Client
 }
 
 func (c *tracingClient) CreateCiBuilderJob(ctx context.Context, params CiBuilderParams) (*batchv1.Job, error) {
