+++ clients/cockroachdb/domain.go
@@ -1,5 +1,7 @@
 package cockroachdb
 
+import "time"
+
 // JobResources represents the used cpu and memory resources for a job and the measured maximum once it's done
 type JobResources struct {
 	CPURequest     float64
@@ -8,4 +10,16 @@ type JobResources struct {
 	MemoryRequest  float64
 	MemoryLimit    float64
 	MemoryMaxUsage float64
+}
+
+// BuildVersionDetail represents a specific build, including version number, repo, branch, revision and manifest
+type BuildVersionDetail struct {
+	ID           int
+	BuildVersion string
+	RepoSource   string
+	RepoFullName string
+	RepoBranch   string
+	RepoRevision string
+	Manifest     string
+	InsertedAt   time.Time
 }
