+++ services/estafette/tracing.go
@@ -11,13 +11,13 @@ import (
 	"github.com/opentracing/opentracing-go/log"
 )
 
-type tracingService struct {
-	Service
-}
-
 // NewTracingService returns a new instance of a tracing Service.
 func NewTracingService(s Service) Service {
 	return &tracingService{s}
+}
+
+type tracingService struct {
+	Service
 }
 
 func (s *tracingService) CreateBuild(ctx context.Context, build contracts.Build, waitForJobToStart bool) (*contracts.Build, error) {
